pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ant-colony"
version = "0.1.0"
description = "Ant colony simulation with pheromone trails and emergent behaviour"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
pygame = ["pygame>=2.5"]
